test(0.10): 73e-T1 桶 D pure-logic 測試搬 tests/unit/（CI 可見）+ 負向合約退役/離線（US6）
把三舊 smoke 檔中從不連網的純邏輯測試搬進 tests/unit/test_scraper_smoke_pure_logic.py，
讓 CI 正確收錄這些確定性守衛（原本藏在 smoke 目錄、CI 看不見）。

- 純搬家組 6 函數：_normalize_fc2_number（去重保留 scrapers.py 5-case 版，含純數字輸入；
  刪 live.py 4-case 版）、JavBus normalize_number、old_api(extract_number)、
  video_model_compat、scraper interface reflection、source_name 唯一性。
- 負向合約組（禁 stub search，CD-73e-7）：
  - AVSOX censored / JavBus invalid 退役（既有 test_avsox_scraper TestEmptySearch/
    TestAllDomainsFail、test_javbus_scraper TestBC8Http404 已等義覆蓋 None-path）。
  - FC2 invalid 搬離線：patch _search_url 回 None 走 None-path（不 stub search）。
- 舊 smoke 只移除上述函數，liveness 測試留給 73e-T2；三舊檔 73e-T7 才整檔刪。

新檔 7 passed 零 skip 零連網；全套（非 smoke/e2e）4229 passed 2 skipped 零回歸。
